Sort simplifications into classic, ht, and intuitionistic
