+++ src/simplifying/fol/mod.rs
@@ -1,2 +1,3 @@
 pub mod classic;
 pub mod ht;
+pub mod intuitionistic;
